# backend/public/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","summary":"User login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"User successfully logged in.","content":{"application/json":{"schema":{"example":{"accessToken":"eyJhbGciOiJIUzI1NiIsInR5....."}}}}}},"tags":["Auth"]}},"/api/v1/posts/{id}":{"get":{"operationId":"PostController_getPostById","summary":"Get post by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Post found","content":{"application/json":{"schema":{"example":{"id":1,"title":"Sample Post Title","detail":"Details of the sample post","userId":1,"communityId":1,"user":{"id":1,"username":"sampleuser","pictureUrl":"https://example.com/user.jpg","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"community":{"id":1,"name":"Sample Community","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"comments":[{"id":1,"comment":"This is a sample comment","userId":1,"postId":1,"createdAt":"2024-08-14T05:00:00.000Z","updatedAt":"2024-08-14T05:00:00.000Z"}],"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"}}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"example":{"message":"Resource not found","statusCode":404}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Posts"]},"put":{"operationId":"PostController_updatePost","summary":"Update post by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDto"}}}},"responses":{"200":{"description":"Post updated successfully","content":{"application/json":{"schema":{"example":{"id":1,"title":"Updated Post Title","detail":"Updated details of the post","userId":1,"communityId":1,"user":{"id":1,"username":"sampleuser","pictureUrl":"https://example.com/user.jpg","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"community":{"id":1,"name":"Sample Community","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"comments":[],"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"example":{"message":"Unauthorized","statusCode":401}}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"example":{"message":"Resource not found","statusCode":404}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Posts"],"security":[{"bearer":[]}]},"delete":{"operationId":"PostController_deletePost","summary":"Delete post by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Post deleted successfully","content":{"application/json":{"schema":{"example":{"id":1,"title":"Sample Post Title","detail":"Details of the sample post","userId":1,"communityId":1,"user":{"id":1,"username":"sampleuser","pictureUrl":"https://example.com/user.jpg","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"community":{"id":1,"name":"Sample Community","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"comments":[],"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"example":{"message":"Unauthorized","statusCode":401}}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"example":{"message":"Resource not found","statusCode":404}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Posts"],"security":[{"bearer":[]}]}},"/api/v1/posts":{"get":{"operationId":"PostController_getPosts","summary":"Get all posts","parameters":[{"name":"fsearch","required":false,"in":"query","description":"Search for titles Case-insensitive","schema":{"example":"healthy","type":"string"}},{"name":"communityId","required":true,"in":"query","description":"ID of the community this post belongs to","schema":{"example":2,"type":"string"}}],"responses":{"200":{"description":"Posts fetched successfully","content":{"application/json":{"schema":{"example":[{"id":1,"title":"Sample Post Title","detail":"Details of the sample post","userId":1,"communityId":1,"user":{"id":1,"username":"sampleuser","pictureUrl":"https://example.com/user.jpg","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"community":{"id":1,"name":"Sample Community","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"comments":[{"id":1,"comment":"This is a sample comment","userId":1,"postId":1,"createdAt":"2024-08-14T05:00:00.000Z","updatedAt":"2024-08-14T05:00:00.000Z"}],"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},{"id":2,"title":"Another Post Title","detail":"Details of another post","userId":2,"communityId":2,"user":{"id":2,"username":"anotheruser","pictureUrl":"https://example.com/user2.jpg","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"community":{"id":2,"name":"Another Community","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"comments":[{"id":2,"comment":"This is another sample comment","userId":2,"postId":2,"createdAt":"2024-08-14T05:00:00.000Z","updatedAt":"2024-08-14T05:00:00.000Z"}],"createdAt":"2024-08-15T04:44:34.066Z","updatedAt":"2024-08-15T04:44:34.066Z"}]}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Posts"]},"post":{"operationId":"PostController_createPost","summary":"Create new post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDto"}}}},"responses":{"201":{"description":"Post created successfully","content":{"application/json":{"schema":{"example":{"id":1,"title":"Sample Post Title","detail":"Details of the sample post","userId":1,"communityId":1,"user":{"id":1,"username":"sampleuser","pictureUrl":"https://example.com/user.jpg","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"community":{"id":1,"name":"Sample Community","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},"comments":[],"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"example":{"message":["field validate failed."],"error":"Bad Request","statusCode":400}}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"example":{"message":"Unauthorized","statusCode":401}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Posts"],"security":[{"bearer":[]}]}},"/api/v1/communities":{"get":{"operationId":"CommunityController_getCommunities","summary":"Get all communities","parameters":[],"responses":{"200":{"description":"Communities fetched successfully","content":{"application/json":{"schema":{"example":[{"id":1,"name":"Healthy Community","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"},{"id":2,"name":"Fitness Community","createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z"}]}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Communities"]}},"/api/v1/comments/{id}":{"get":{"operationId":"CommentController_getCommentById","summary":"Get comment by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Comment found","content":{"application/json":{"schema":{"example":{"id":1,"comment":"This is a comment","userId":1,"postId":1,"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z","user":{"id":1,"username":"testuser","pictureUrl":"https://example.com/user.jpg"}}}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"example":{"message":"Resource not found","statusCode":404}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Comments"]},"put":{"operationId":"CommentController_updateComment","summary":"Update comment by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCommentDto"}}}},"responses":{"200":{"description":"Comment updated successfully","content":{"application/json":{"schema":{"example":{"id":1,"comment":"Updated comment","userId":1,"postId":1,"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-15T04:44:34.066Z","user":{"id":1,"username":"testuser","pictureUrl":"https://example.com/user.jpg"}}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"example":{"message":["field validate failed."],"error":"Bad Request","statusCode":400}}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"example":{"message":"Resource not found","statusCode":404}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Comments"],"security":[{"bearer":[]}]},"delete":{"operationId":"CommentController_deleteComment","summary":"Delete comment by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Comment deleted successfully","content":{"application/json":{"schema":{"example":{"id":1,"comment":"This is a comment","userId":1,"postId":1,"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z","user":{"id":1,"username":"testuser","pictureUrl":"https://example.com/user.jpg"}}}}}},"404":{"description":"Not found","content":{"application/json":{"schema":{"example":{"message":"Resource not found","statusCode":404}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Comments"],"security":[{"bearer":[]}]}},"/api/v1/comments":{"post":{"operationId":"CommentController_createComment","summary":"Create new comment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentDto"}}}},"responses":{"201":{"description":"Comment created successfully","content":{"application/json":{"schema":{"example":{"id":1,"comment":"This is a comment","userId":1,"postId":1,"createdAt":"2024-08-14T04:44:34.066Z","updatedAt":"2024-08-14T04:44:34.066Z","user":{"id":1,"username":"testuser","pictureUrl":"https://example.com/user.jpg"}}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"example":{"message":["field validate failed."],"error":"Bad Request","statusCode":400}}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"example":{"message":"Internal server error","error":"Internal Server Error","statusCode":500}}}}}},"tags":["Comments"],"security":[{"bearer":[]}]}}},"info":{"title":"DataWowTest API","description":"DataWowTest Swagger Open API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string","example":"user1"}},"required":["username"]},"CreatePostDto":{"type":"object","properties":{"title":{"type":"string","example":"New Post Title","description":"Title of the post"},"detail":{"type":"string","example":"Post details go here...","description":"Details of the post"},"communityId":{"type":"number","example":1,"description":"ID of the community this post belongs to"}},"required":["title","detail","communityId"]},"UpdatePostDto":{"type":"object","properties":{"title":{"type":"string","example":"New Post Title","description":"Title of the post"},"detail":{"type":"string","example":"Post details go here...","description":"Details of the post"},"communityId":{"type":"number","example":1,"description":"ID of the community this post belongs to"}}},"CreateCommentDto":{"type":"object","properties":{"comment":{"type":"string","example":"This is a comment","description":"The content of the comment"},"postId":{"type":"number","example":1,"description":"The ID of the post this comment belongs to"}},"required":["comment","postId"]},"UpdateCommentDto":{"type":"object","properties":{"comment":{"type":"string","example":"This is a comment","description":"The content of the comment"},"postId":{"type":"number","example":1,"description":"The ID of the post this comment belongs to"}}}}}}